Checking Code for the Mean Values
Jupyter recorded some changes that I have made to the document and wants me to stage and push the changes to Git Hub. Changes that I have made before have been reverted.

In the "Wrangling and Cleaning The Data" section, I was experimenting with the ways to find the mean values of numerical variables in our data set to discuss the effects of preprocessing. Summarize do the same thing as mutate, select and slice altogether.

I have double checked to make sure the code was correct and kept the mutate, select and slice altogether structure the same. To sum up, nothing has been changed before. All changes have been reverted.